# [perf] Accelerate _inside_class()