--- a/python/taichi/lang/kernel_impl.py
+++ b/python/taichi/lang/kernel_impl.py
@@ -643,9 +643,8 @@ _KERNEL_CLASS_STACKFRAME_STMT_RES = [
 def _inside_class(level_of_class_stackframe):
     try:
         maybe_class_frame = inspect.currentframe()
-        while level_of_class_stackframe > 0:
+        for _ in range(level_of_class_stackframe):
             maybe_class_frame = maybe_class_frame.f_back
-            level_of_class_stackframe -= 1
         statement_list = inspect.getframeinfo(maybe_class_frame)[3]
         first_statment = statement_list[0].strip()
         for pat in _KERNEL_CLASS_STACKFRAME_STMT_RES: